Cleaned up code and done some basic optimisations.
+ Many more function parameters are now const.
+ Setting of collision flags looks a *lot* nicer now.

+++ src/Player.hpp
@@ -35,7 +35,7 @@ public:
 
     Weapon weapon;
 
-    map<string, unsigned short int> stats {
+    map<string, int8_t> stats {
         {"hea", 100},
         {"atk", 0},
         {"def", 0},
